# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Hypermedia project by Roccamena, Romanò, Buttolo"
  version: "1.0.0"
  title: "Child4Help"
  termsOfService: "http://swagger.io/terms/"
host: "localhost:8080"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
...
  /events/items:
    get:
      tags:
      - "event"
      summary: "Get all events items"
      operationId: "getEventsItemAll"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Events page number to get"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EventItem"
      x-swagger-router-controller: "Event"
  /events/items/{month}:
    get:
      tags:
      - "event"
      summary: "Get all events by month"
      operationId: "getEventByMonth"
      produces:
      - "application/json"
      parameters:
      - name: "month"
        in: "path"
        description: "Month of events to return"
        required: true
        type: "integer"
        maximum: 12
        minimum: 1
        format: "int64"
      - name: "page"
        in: "query"
        description: "Events page number to get"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EventItem"
        "400":
          description: "Invalid month"
      x-swagger-router-controller: "Event"
  /event/{id_event}:
    get:
      tags:
      - "event"
      summary: "Get an event by ID"
      operationId: "getEventByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_event"
        in: "path"
        description: "ID of event to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/Event"
        "400":
          description: "Invalid ID"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
    delete:
      tags:
      - "event"
      summary: "Delete an event by ID"
      operationId: "deleteEvent"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_event"
        in: "path"
        description: "ID of event to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid ID"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /event/{id_event}/page1:
    get:
      tags:
      - "event"
      summary: "Get an event page 1 data by ID"
      operationId: "getEventPage1ByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_event"
        in: "path"
        description: "ID of event to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/EventPage1"
        "400":
          description: "Invalid ID"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /event/{id_event}/page2:
    get:
      tags:
      - "event"
      summary: "Get an event page 2 data by ID"
      operationId: "getEventPage2ByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_event"
        in: "path"
        description: "ID of event to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/EventPage2"
        "400":
          description: "Invalid ID"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /event:
    post:
      tags:
      - "event"
      summary: "Add a new event"
      operationId: "addEvent"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "event"
        description: "Event object to be added"
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
      x-swagger-router-controller: "Event"
    put:
      tags:
      - "event"
      summary: "Update an event"
      operationId: "updateEvent"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "event"
        description: "Event object to be updated"
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        "200":
          description: "Successful operation"
        "400":
          description: "Invalid input"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /event/{id_event}/related:
    get:
      tags:
      - "event"
      summary: "Get related events of a selected event by ID"
      operationId: "getRelatedEventsByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_event"
        in: "path"
        description: "ID of event to search for related"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EventItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /event/{id_event}/services:
    get:
      tags:
      - "event"
      summary: "Get services of a selected event by ID"
      operationId: "getEventServicesByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_event"
        in: "path"
        description: "ID of event to search for services"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /event/{id_event}/person:
    get:
      tags:
      - "event"
      summary: "Get person of a selected event by ID"
      operationId: "getEventPersonByID"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id_event"
        in: "path"
        description: "ID of event to search for person"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            $ref: "#/definitions/PersonItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
  /event/{id_event}/articles:
    get:
      tags:
      - "event"
      summary: "Get articles of a selected event by ID"
      operationId: "getEventArticlesByID"
      produces:
      - "application/json"
      parameters:
      - name: "id_event"
        in: "path"
        description: "ID of event to search for articles"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ArticleItem"
        "400":
          description: "Invalid ID"
        "404":
          description: "Event not found"
      x-swagger-router-controller: "Event"
...
definitions:
...
  ServiceItem:
    type: "object"
    required:
    - "id_service"
    properties:
      id_service:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      presentation:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "ServiceItem"
    example:
      presentation: "presentation"
      name: "name"
      photo_url: "photo_url"
      id_service: 0
  Event:
    type: "object"
    required:
    - "id_event"
    properties:
      id_event:
        type: "integer"
        format: "int64"
      id_person:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      date:
        type: "string"
        format: "date"
      location:
        type: "string"
      description:
        type: "string"
      pract_info:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "Event"
    example:
      date: "2000-01-23"
      id_person: 6
      name: "name"
      description: "description"
      pract_info: "pract_info"
      location: "location"
      id_event: 0
      photo_url: "photo_url"
  EventPage1:
    type: "object"
    required:
    - "id_event"
    properties:
      id_event:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      description:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "EventPage1"
    example:
      name: "name"
      description: "description"
      id_event: 0
      photo_url: "photo_url"
  EventPage2:
    type: "object"
    required:
    - "id_event"
    properties:
      id_event:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      date:
        type: "string"
        format: "date"
      location:
        type: "string"
      pract_info:
        type: "string"
    xml:
      name: "EventPage2"
    example:
      date: "2000-01-23"
      name: "name"
      pract_info: "pract_info"
      location: "location"
      id_event: 0
  EventItem:
    type: "object"
    required:
    - "id_event"
    properties:
      id_event:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      date:
        type: "string"
        format: "date"
      location:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "EventItem"
    example:
      date: "2000-01-23"
      name: "name"
      location: "location"
      id_event: 0
      photo_url: "photo_url"
...
  PersonItem:
    type: "object"
    required:
    - "id_person"
    properties:
      id_person:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "PersonItem"
    example:
      id_person: 0
      name: "name"
      photo_url: "photo_url"
...
  ArticleItem:
    type: "object"
    required:
    - "id_article"
    properties:
      id_article:
        type: "integer"
        format: "int64"
      author:
        type: "string"
      publication_date:
        type: "string"
        format: "date"
      title:
        type: "string"
      photo_url:
        type: "string"
    xml:
      name: "ArticleItem"
    example:
      id_article: 0
      author: "author"
      publication_date: "2000-01-23"
      photo_url: "photo_url"
      title: "title"
...